+++ npm-packages/react-api/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @pulse-editor/react-api
 
+## 0.1.1-alpha.41
+
+### Patch Changes
+
+- Update register action hook for react api
+- Updated dependencies
+  - @pulse-editor/shared-utils@0.1.1-alpha.41
+
+## 0.1.1-alpha.40
+
+### Patch Changes
+
+- Rename pre-regsitered actions
+- Updated dependencies
+  - @pulse-editor/shared-utils@0.1.1-alpha.40
+
 ## 0.1.1-alpha.39
 
 ### Patch Changes